backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.0
tensorflow==2.13.0
huggingface-hub==0.17.3
numpy==1.24.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
Flask-Mail==0.9.1
APScheduler==3.10.4